pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-solar-manager"
version = "0.1.0"
description = "Home Assistant integration for the Solar Manager local API."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "homeassistant>=2026.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["custom_components"]

[tool.uv]
override-dependencies = ["lru-dict>=1.4.1"]

[tool.pyright]
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pyright>=1.1.396",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
]
